Make `src` an optional paramter
I'm using this to create a sandboxed iFrame that shouldn't have a `src`. Not passing a `src` makes the iframe include `src="undefined"`. Passing it a blank `src` makes the iframe include `src(unknown)`. Both these options are conflicting with my ability to inject content into it. I think the best thing to do is simply not set the attribute if it's not provided to the constructor. 

Happy to talk some more about my (admittedly a-typical)  use case  if you'd like.